requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.74
matplotlib==3.7.1
Pillow==10.0.0
gradio==3.41.2
tqdm==4.65.0
scikit-learn==1.3.0
pandas==2.0.3
seaborn==0.12.2